moved instruments out of generation
