+++ src/export/arrangement_generator.py
@@ -5,7 +5,8 @@ from typing import Dict, Type
 from common.note_collection import NoteCollection
 
 from generation.chord_progression import ChordProgression
-import generation.instruments.base as instrument
+
+import instruments.base as instrument
 
 from export.arrangement import Arrangement, ArrangementMetadata
