+++ src/main.py
@@ -11,10 +11,10 @@ from melody_extraction.signal import SignalMelodyExtractor
 
 from generation.chord_progression import ChordProgression
 from generation.viterbi import ViterbiChordProgressionGenerator
-from generation.instruments.voice import Voice
-from generation.instruments.piano import Piano
-from generation.instruments.bass_drum import BassDrum
-from generation.instruments.snare_drum import SnareDrum
+from instruments.voice import Voice
+from instruments.piano import Piano
+from instruments.bass_drum import BassDrum
+from instruments.snare_drum import SnareDrum
 
 from export.arrangement import Arrangement, ArrangementExportConfig, ArrangementMetadata
 from export.arrangement_generator import ArrangementGenerator
